+++ google2/src/main/scala/org/broadinstitute/dsde/workbench/google2/GKEInterpreter.scala
@@ -4,7 +4,15 @@ import cats.effect.concurrent.Semaphore
 import cats.effect.{Async, Blocker, ContextShift, Timer}
 import cats.mtl.ApplicativeAsk
 import com.google.cloud.container.v1.ClusterManagerClient
-import com.google.container.v1.{Cluster, CreateClusterRequest, GetOperationRequest, IPAllocationPolicy, Operation}
+import com.google.container.v1.{
+  Cluster,
+  CreateClusterRequest,
+  CreateNodePoolRequest,
+  GetOperationRequest,
+  IPAllocationPolicy,
+  NodePool,
+  Operation
+}
 import fs2.Stream
 import io.chrisdavenport.log4cats.StructuredLogger
 import org.broadinstitute.dsde.workbench.{DoneCheckable, RetryConfig}
@@ -14,7 +22,7 @@ import org.broadinstitute.dsde.workbench.model.TraceId
 
 import scala.concurrent.duration.FiniteDuration
 
-class GKEInterpreter[F[_]: Async: StructuredLogger: Timer: ContextShift](
+final class GKEInterpreter[F[_]: Async: StructuredLogger: Timer: ContextShift](
   clusterManagerClient: ClusterManagerClient,
   blocker: Blocker,
   blockerBound: Semaphore[F],
@@ -22,15 +30,15 @@ class GKEInterpreter[F[_]: Async: StructuredLogger: Timer: ContextShift](
 ) extends GKEService[F] {
 
   override def createCluster(
-    kubernetesClusterRequest: KubernetesCreateClusterRequest
+    request: KubernetesCreateClusterRequest
   )(implicit ev: ApplicativeAsk[F, TraceId]): F[Operation] = {
-    val parent = Parent(kubernetesClusterRequest.project, kubernetesClusterRequest.location)
+    val parent = Parent(request.project, request.location)
 
     val createClusterRequest: CreateClusterRequest = CreateClusterRequest
       .newBuilder()
-      .setParent(parent.parentString)
+      .setParent(parent.toString)
       .setCluster(
-        kubernetesClusterRequest.cluster.toBuilder
+        request.cluster.toBuilder
           .setIpAllocationPolicy( //otherwise it uses the legacy one, which is insecure. See https://cloud.google.com/kubernetes-engine/docs/how-to/alias-ips
             IPAllocationPolicy
               .newBuilder()
@@ -41,23 +49,50 @@ class GKEInterpreter[F[_]: Async: StructuredLogger: Timer: ContextShift](
 
     tracedGoogleRetryWithBlocker(
       Async[F].delay(clusterManagerClient.createCluster(createClusterRequest)),
-      f"com.google.cloud.container.v1.ClusterManagerClient.createCluster(${kubernetesClusterRequest})"
+      f"com.google.cloud.container.v1.ClusterManagerClient.createCluster(${request})"
     )
   }
 
   override def getCluster(clusterId: KubernetesClusterId)(implicit ev: ApplicativeAsk[F, TraceId]): F[Option[Cluster]] =
     tracedGoogleRetryWithBlocker(
       recoverF(
-        Async[F].delay(clusterManagerClient.getCluster(clusterId.idString)),
+        Async[F].delay(clusterManagerClient.getCluster(clusterId.toString)),
         whenStatusCode(404)
       ),
-      f"com.google.cloud.container.v1.ClusterManagerClient.getCluster(${clusterId.idString})"
+      f"com.google.cloud.container.v1.ClusterManagerClient.getCluster(${clusterId.toString})"
     )
 
   override def deleteCluster(clusterId: KubernetesClusterId)(implicit ev: ApplicativeAsk[F, TraceId]): F[Operation] =
     tracedGoogleRetryWithBlocker(
-      Async[F].delay(clusterManagerClient.deleteCluster(clusterId.idString)),
-      f"com.google.cloud.container.v1.ClusterManagerClient.deleteCluster(${clusterId.idString})"
+      Async[F].delay(clusterManagerClient.deleteCluster(clusterId.toString)),
+      f"com.google.cloud.container.v1.ClusterManagerClient.deleteCluster(${clusterId.toString})"
+    )
+
+  override def createNodepool(
+    request: KubernetesCreateNodepoolRequest
+  )(implicit ev: ApplicativeAsk[F, TraceId]): F[Operation] = {
+    val createNodepoolRequest: CreateNodePoolRequest = CreateNodePoolRequest
+      .newBuilder()
+      .setParent(request.clusterId.toString)
+      .setNodePool(request.nodepool.toBuilder)
+      .build()
+
+    tracedGoogleRetryWithBlocker(
+      Async[F].delay(clusterManagerClient.createNodePool(createNodepoolRequest)),
+      f"com.google.cloud.container.v1.ClusterManagerClient.createNodepool(${request})"
+    )
+  }
+
+  override def getNodepool(nodepoolId: NodepoolId)(implicit ev: ApplicativeAsk[F, TraceId]): F[NodePool] =
+    tracedGoogleRetryWithBlocker(
+      Async[F].delay(clusterManagerClient.getNodePool(nodepoolId.toString)),
+      f"com.google.cloud.container.v1.ClusterManagerClient.getNodepool(${nodepoolId.toString})"
+    )
+
+  override def deleteNodepool(nodepoolId: NodepoolId)(implicit ev: ApplicativeAsk[F, TraceId]): F[Operation] =
+    tracedGoogleRetryWithBlocker(
+      Async[F].delay(clusterManagerClient.deleteNodePool(nodepoolId.toString)),
+      f"com.google.cloud.container.v1.ClusterManagerClient.deleteNodepool(${nodepoolId.toString})"
     )
 
   //delete and create operations take around ~5mins with simple tests, could be longer for larger clusters